First draft of canon support
Former-commit-id: 75c43d058412774b9017c9a603105ccaa80f7641 [formerly 469fdbb4210c532ca8e67bb7df3d1fdc16386efa]
Former-commit-id: 796d2d510b1a64419999a3ee7fa4082580e84fc0 [formerly 3b2f8a274cbc7e48974895ce374d808c693eae4a]
Former-commit-id: 99fb4927a24c15ce5dd8d1dcc4042f09f4333653
Former-commit-id: 5539dccc3d0ef45d82eb9ee6d6b6c9ff40c04115

+++ Source/Raw.Net/Source/Format/Tiff/IFD/IFD.cs
@@ -12,7 +12,7 @@ namespace RawNet
         public ushort tagNumber;
         public Dictionary<TagType, Tag> tags;
         public List<IFD> subIFD = new List<IFD>();
-        public ushort nextOffset { get; set; }
+        public uint nextOffset { get; set; }
         public Endianness endian;
         public int depth;
 
@@ -227,7 +227,7 @@ namespace RawNet
                     Debug.WriteLine("tags already exist");
                 }
             }
-            nextOffset = fileStream.ReadUInt16();
+            nextOffset = fileStream.ReadUInt32();
         }
 
         internal bool hasEntry(TagType t)
